periodic update filters only relevant script hashes
