+++ src/bin/electrs.rs
@@ -8,25 +8,21 @@ use error_chain::ChainedError;
 use std::process;
 use std::sync::Arc;
 use std::time::Duration;
-use std::collections::HashSet;
-use bitcoin_hashes::sha256d::Hash as Sha256dHash;
-use bitcoin_hashes::Hash;
 
 use electrs::{
     app::App,
     bulk,
     cache::{BlockTxIDsCache, TransactionCache},
     config::Config,
-    daemon::{Daemon, tx_from_value},
+    daemon::Daemon,
     errors::*,
-    index::{Index, compute_script_hash},
+    index::Index,
     metrics::Metrics,
     query::Query,
     rpc::RPC,
     signal::Waiter,
     store::{full_compaction, is_fully_compacted, DBStore},
 };
-use bitcoin::Transaction;
 
 fn run_server(config: &Config) -> Result<()> {
     let signal = Waiter::start();
@@ -69,95 +65,19 @@ fn run_server(config: &Config) -> Result<()> {
     loop {
         let new_block_headers = app.update(&signal)?;
         let new_mempool_txs = query.update_mempool()?;
-        let script_hashes = get_script_hashes_in_blocks(app.daemon(), &query,new_block_headers)?
-            .extend(get_script_hashes_in_mempool_txs(app.daemon(), &query, new_mempool_txs)?);
+        let mut script_hashes = query.get_script_hashes_in_blocks(new_block_headers)?;
+        script_hashes.extend(query.get_script_hashes_in_mempool_txs(new_mempool_txs)?);
+
         server
             .get_or_insert_with(|| RPC::start(config.electrum_rpc_addr, query.clone(), &metrics))
-            .notify(); // update subscribed clients
+            .notify(script_hashes); // update subscribed clients
+
         if let Err(err) = signal.wait(Duration::from_secs(5)) {
             info!("stopping server: {}", err);
             break;
         }
     }
     Ok(())
-}
-
-
-fn get_script_hashes_in_blocks(daemon: &Daemon, query: &Arc<Query>, block_hashes: Vec<Sha256dHash>) -> Result<HashSet<Sha256dHash>> {
-    let mut script_hashes = HashSet::<Sha256dHash>::new();
-    let blocks = daemon.getblocks(block_hashes.as_ref())?;
-    for block in blocks {
-        for tx in block.txdata {
-            if tx.is_coin_base() {
-                continue;
-            }
-
-            // for each script_hash in tx.inputs
-            for input in tx.input.iter() {
-                // find output, get the relevant script hash in it
-                let previous_output_tx = tx_from_value(
-                    query.get_transaction(&input.previous_output.txid, false)?)?;
-                let previous_output = previous_output_tx.output.get(input.previous_output.vout as usize)
-                    .chain_err(|| format!("failed finding previous output {}:{}", input.previous_output.txid, input.previous_output.vout))?;
-                let script_hash =
-                    Sha256dHash::from_slice(&compute_script_hash(&previous_output.script_pubkey[..]))
-                        .expect("failed computing script hash for output.script_pubkey");
-                // if script_hash in self.script_hashes, add to the relevant script hashes list
-                script_hashes.insert(script_hash);
-            }
-
-            for (i, output) in tx.output.iter().enumerate() {
-                let script_hash =
-                    Sha256dHash::from_slice(&compute_script_hash(&output.script_pubkey[..]))
-                        .expect(&format!("failed computing script hash for output {}:{}", tx.txid(), i));
-                // if script_hash in self.script_hashes:
-                script_hashes.insert(script_hash);
-            }
-        }
-    }
-
-    Ok(script_hashes)
-}
-
-fn get_script_hashes_in_mempool_txs(daemon: &Daemon, query: &Arc<Query>, txs: Vec<Transaction>) -> Result<HashSet<Sha256dHash>> {
-    let mut script_hashes: HashSet<Sha256dHash> = HashSet::<Sha256dHash>::new();
-    for tx in txs {
-        for input in tx.input.iter() {
-            // find output, get the relevant script hash in it
-            let previous_output = daemon.get_confirmed_utxo(&input.previous_output.txid, input.previous_output.vout)
-                .or_else(|_e| {
-                    // possibly failed because previous output's transaction itself is unconfirmed, so search in mempool.
-                    debug!("failed to find a confirmed previous output {}:{}", &input.previous_output.txid, &input.previous_output.vout);
-                    let previous_output_tx: Transaction = query.tracker().read().unwrap()
-                        .get_txn(&input.previous_output.txid)
-                        .chain_err(|| format!("failed to find unconfirmed previous output tx {}:{}",
-                            &input.previous_output.txid, &input.previous_output.vout))?;
-
-                    previous_output_tx.output.get(input.previous_output.vout as usize)
-                        .map(|output| output.clone())
-                        .chain_err(|| format!("failed to find previous output index in tx {}:{}",
-                            &input.previous_output.txid, &input.previous_output.vout))
-                })
-                .expect(&format!("failed to get previous output {}:{}", &input.previous_output.txid, &input.previous_output.vout));
-
-            let script_hash =
-                Sha256dHash::from_slice(&compute_script_hash(&previous_output.script_pubkey[..]))
-                    .expect(&format!("failed computing script hash of previous output {}:{}",
-                        &input.previous_output.txid, &input.previous_output.vout));
-
-            script_hashes.insert(script_hash);
-        }
-
-        for (i, output) in tx.output.iter().enumerate() {
-            let script_hash =
-                Sha256dHash::from_slice(&compute_script_hash(&output.script_pubkey[..]))
-                    .expect(&format!("failed computing script hash for output {}:{}", tx.txid(), i));
-
-            script_hashes.insert(script_hash);
-        }
-    }
-
-    Ok(script_hashes)
 }
 
 fn main() {
