+++ src/query.rs
@@ -6,10 +6,11 @@ use bitcoin_hashes::Hash;
 use crypto::digest::Digest;
 use crypto::sha2::Sha256;
 use serde_json::Value;
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 use std::sync::{Arc, RwLock};
 
 use crate::app::App;
+use crate::daemon::tx_from_value;
 use crate::cache::TransactionCache;
 use crate::errors::*;
 use crate::index::{compute_script_hash, TxInRow, TxOutRow, TxRow};
@@ -521,7 +522,80 @@ impl Query {
         self.app.get_banner()
     }
 
-    pub fn tracker(&self) -> &RwLock<Tracker> {
-        &self.tracker
+    pub fn get_script_hashes_in_blocks(&self, block_hashes: Vec<Sha256dHash>) -> Result<HashSet<Sha256dHash>> {
+        let mut script_hashes = HashSet::<Sha256dHash>::new();
+        let blocks = self.app.daemon().getblocks(block_hashes.as_ref())?;
+        for block in blocks {
+            for tx in block.txdata {
+                if tx.is_coin_base() {
+                    continue;
+                }
+
+                // for each script_hash in tx.inputs
+                for input in tx.input.iter() {
+                    // find output, get the relevant script hash in it
+                    let previous_output_tx = tx_from_value(
+                        self.get_transaction(&input.previous_output.txid, false)?)?;
+                    let previous_output = previous_output_tx.output.get(input.previous_output.vout as usize)
+                        .chain_err(|| format!("failed finding previous output {}:{}", input.previous_output.txid, input.previous_output.vout))?;
+                    let script_hash =
+                        Sha256dHash::from_slice(&compute_script_hash(&previous_output.script_pubkey[..]))
+                            .expect("failed computing script hash for output.script_pubkey");
+                    // if script_hash in self.script_hashes, add to the relevant script hashes list
+                    script_hashes.insert(script_hash);
+                }
+
+                for (i, output) in tx.output.iter().enumerate() {
+                    let script_hash =
+                        Sha256dHash::from_slice(&compute_script_hash(&output.script_pubkey[..]))
+                            .expect(&format!("failed computing script hash for output {}:{}", tx.txid(), i));
+                    // if script_hash in self.script_hashes:
+                    script_hashes.insert(script_hash);
+                }
+            }
+        }
+
+        Ok(script_hashes)
+    }
+
+    pub fn get_script_hashes_in_mempool_txs(&self, txs: Vec<Transaction>) -> Result<HashSet<Sha256dHash>> {
+        let mut script_hashes = HashSet::<Sha256dHash>::new();
+        for tx in txs {
+            for input in tx.input.iter() {
+                // find output, get the relevant script hash in it
+                let previous_output = self.app.daemon().get_confirmed_utxo(&input.previous_output.txid, input.previous_output.vout)
+                    .or_else(|_e| {
+                        // possibly failed because previous output's transaction itself is unconfirmed, so search in mempool.
+                        debug!("failed to find a confirmed previous output {}:{}", &input.previous_output.txid, &input.previous_output.vout);
+                        let previous_output_tx = self.tracker.read().unwrap()
+                            .get_txn(&input.previous_output.txid)
+                            .chain_err(|| format!("failed to find unconfirmed previous output tx {}:{}",
+                                                  &input.previous_output.txid, &input.previous_output.vout))?;
+
+                        previous_output_tx.output.get(input.previous_output.vout as usize)
+                            .map(|output| output.clone())
+                            .chain_err(|| format!("failed to find previous output index in tx {}:{}",
+                                                  &input.previous_output.txid, &input.previous_output.vout))
+                    })
+                    .expect(&format!("failed to get previous output {}:{}", &input.previous_output.txid, &input.previous_output.vout));
+
+                let script_hash =
+                    Sha256dHash::from_slice(&compute_script_hash(&previous_output.script_pubkey[..]))
+                        .expect(&format!("failed computing script hash of previous output {}:{}",
+                                         &input.previous_output.txid, &input.previous_output.vout));
+
+                script_hashes.insert(script_hash);
+            }
+
+            for (i, output) in tx.output.iter().enumerate() {
+                let script_hash =
+                    Sha256dHash::from_slice(&compute_script_hash(&output.script_pubkey[..]))
+                        .expect(&format!("failed computing script hash for output {}:{}", tx.txid(), i));
+
+                script_hashes.insert(script_hash);
+            }
+        }
+
+        Ok(script_hashes)
     }
 }
